% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Fed_MPAs_clipped}
\alias{Fed_MPAs_clipped}
\title{Example shape file for Federal MPA polygons on West Coast.}
\format{
An spatal feature file
}
\usage{
Fed_MPAs_clipped
}
\description{
Example shape file for Federal MPA polygons on West Coast.
}
\keyword{datasets}
