% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterSample.r
\name{getIndividualBoxIndices}
\alias{getIndividualBoxIndices}
\title{This is an internal Master Sample function to assign
indices to points based on the discrete Halton Box overlay.}
\usage{
getIndividualBoxIndices(pts, J = NULL, bb)
}
\arguments{
\item{J}{Integer for number of Halton Boxes to make. If not set it defaults to 100m roughly.}

\item{bb}{Master Sample bounding box.}

\item{input}{An sp or sf spatial points. Accepts either. Or even a data frame with X, Y names.}

\item{base}{Generally 2,3. If you change it read the literature.}

\item{seed}{Master Sample Seed for two bases.}

\item{s1}{Permutation of x orderings (0, 1)}

\item{s2}{Permutation of y orderings (0, 1, 2)}
}
\description{
This is an internal Master Sample function to assign
indices to points based on the discrete Halton Box overlay.
}
