% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterSample.r
\name{getProj}
\alias{getProj}
\title{Define spatial objects in projection of the master sample.}
\usage{
getProj()
}
\description{
Default projection of the master sample. Needed for consistency for the entire bounding box.
}
