% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defintions.r
\name{getSeed}
\alias{getSeed}
\title{Random seed definition for Western Canada Marine Master Sample.}
\usage{
getSeed()
}
\description{
Defines the random seed specific to the Western Canada Marine Master Sample.
}
