% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaltonFunctions.r
\name{makeFrame}
\alias{makeFrame}
\title{Create a Halton Frame raster, defined in sf based on the bounding box master sample.}
\usage{
makeFrame(base = c(2, 3), J = c(2, 2), bb, rotate = FALSE)
}
\arguments{
\item{base}{Co-prime base for BAS, do not change from 2,3.}

\item{J}{Definition for the number of grid cells of Halton frame.}

\item{bb}{Bounding box shapefile with centroid, random seed, rotation.}

\item{rotate}{Boolean if you want to rotate shape before exporting.}
}
\value{
rotated sf spatial object.
}
\description{
Make a Halton Frame based on B = 2^J[1]*3^J[2] grid cells. If rotation is required, will return rotated.
This function is an internal function simply to select sub BAS points without having to do spatial clipping at the point level.
}
\examples{
\dontrun{
bb <- getBB()
haltonFrame <- makeFrame(J = c(8,4), bb = bb)
}
}
