% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterSample.r
\name{masterSampleSelect}
\alias{masterSampleSelect}
\title{Generate sample points using BAS master sample}
\usage{
masterSampleSelect(shp, N = 100, bb = NULL, nExtra = 5000, printJ = FALSE)
}
\description{
Generates BAS sample points in a specified sample frame based on New Zealand terrestrial mastersample. Users need to specify 'island' the sample frame is in and how many points are required.
}
