% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{NS_bioregion}
\alias{NS_bioregion}
\title{Bio region information for stratification}
\format{
An spatal feature file
}
\usage{
NS_bioregion
}
\description{
Bio region information for stratification
}
\keyword{datasets}
