% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.r
\name{rot}
\alias{rot}
\title{Rotation Matrix}
\usage{
rot(a)
}
\arguments{
\item{a}{radians of rotation.}
}
\value{
Matrix
}
\description{
Generate a rotation matrix for rotating objects later
}
\examples{
\dontrun{
rot(pi)
}
}
