% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.r
\name{rotate.shp}
\alias{rotate.shp}
\title{Rotate a polygon around the centroid of a Master Sample bounding box.}
\usage{
rotate.shp(shp, bb, back = TRUE)
}
\arguments{
\item{shp}{Any polygon within the sample frame defined as a spatial features object.}

\item{bb}{Special shape file defining the bounding box with attributes for centroid and rotation.}

\item{back}{Boolean for whether or not the rotation is back to the original rotated bounding box.}
}
\value{
rotated sf spatial object.
}
\description{
Given some polygon within the bounding box of a Master Sample rotate it by theta defined
by that bounding box either backwards or forwards.
}
\examples{
\dontrun{
data(NS_bioregion)
bb <- getBB()
pts.shp <- rotate.shp(shp = NS_bioregion, bb = bb)
}
}
