% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaltonFunctions.r
\name{shape2Frame}
\alias{shape2Frame}
\title{Clip a Halton Frame based on the bounding box to the current shape.}
\usage{
shape2Frame(
  shp,
  bb = NULL,
  base = c(2, 3),
  J = c(2, 2),
  projstring = NULL,
  rotate = FALSE
)
}
\arguments{
\item{shp}{shape as spatial features object to wrap into Halton frame.}

\item{bb}{Master Sample bounding box.}

\item{base}{Co-prime base for BAS, do not change from 2,3.}

\item{J}{Definition for the number of grid cells of Halton frame.}

\item{projstring}{Projection that the master sample is in, can be passed as part of the bounding box.}

\item{rotate}{Boolean if you want to rotate shape before exporting.}
}
\value{
rotated sf spatial object.
}
\description{
Take a shapefile as a sf object and clips boxes from a Halton frame around it. Size of those boxes is chosen
by choosing J, the number of base 2,3 powers to subdivide. Intended for internal use but can be useful in other
context.
}
\examples{
\dontrun{
bb <- getBB()
data(NS_bioregion)
haltonBoxes <- shape2Frame(shp = NS_biogregion, J = c(6,4), bb = bb)
}
}
